refactor(fibonacci): move pythonkit -> mathkit
